Begun working on bounce-lighting
There are still bugs, seemingly relating to the colour of the bounced light. This might be related to the way I'm reading albedo textures.

It seems to be sampling the lightmap information perfectly well, but not the albedo texture. I'll have to look into this when I get the chance